changed name componenet productPage to productInfo addedprops title in component mostViewed
