+++ goldmarket/src/components/productInfo/aboutProduct/productContactThe.tsx
@@ -1,0 +1,15 @@
+import React, {Component} from 'react';
+
+class ProductContactThe extends Component {
+	render() {
+		return (
+			<div className={'fines  text-center'}>
+				<span>
+						Contact the seller to change the size and weight of this jewelery
+				</span>
+			</div>
+		);
+	}
+}
+
+export {ProductContactThe};
