+++ goldmarket/src/pages/home/index.tsx
@@ -14,7 +14,7 @@ import NewCollection from 'src/components/newCollections';
 import SellWithUs from '../../components/information/sellWithUsBanner';
 import WhyGoldCenter from '../../components/information/whyGoldCenterBanner';
 import Video from '../../components/header/video';
-import ProductPage from '../../components/productPage';
+import ProductPage from '../../components/productInfo';
 import AddProduct from 'src/components/product/AddProduct';
 import SignUp from '../../components/signUp';
